+++ packages/core/src/lib/wait_for_remote_peer.ts
@@ -10,7 +10,7 @@ const log = debug("waku:wait-for-remote-peer");
  * Wait for a remote peer to be ready given the passed protocols.
  * Must be used after attempting to connect to nodes, using
  * {@link @waku/core.WakuNode.dial} or a bootstrap method with
- * {@link @waku/create.createLightNode}.
+ * {@link @waku/sdk.createLightNode}.
  *
  * If the passed protocols is a GossipSub protocol, then it resolves only once
  * a peer is in a mesh, to help ensure that other peers will send and receive
