+++ packages/tests/tests/filter_v2.node.spec.ts
@@ -5,13 +5,13 @@ import {
   DefaultPubSubTopic,
   waitForRemotePeer,
 } from "@waku/core";
-import { createLightNode } from "@waku/create";
 import type {
   IFilterV2,
   IFilterV2Subscription,
   LightNode,
 } from "@waku/interfaces";
 import { Protocols } from "@waku/interfaces";
+import { createLightNode } from "@waku/sdk";
 import { bytesToUtf8, utf8ToBytes } from "@waku/utils/bytes";
 import { expect } from "chai";
 import debug from "debug";
